requirements.txt
gradio==4.26.0
pandas==2.1.0
openai==1.12.0
lxml==5.2.1
python-dotenv==1.0.1